Move cell_content_area_offset 0 translation to cursor layer.
cell_content_area_offset 0 is translated into MAX_PAGE_SIZE. However
validating cell_content_area_offset 0 in btree parsing layer is not
efficient because it causes conditional branch.

As allocateSpace() in SQLite btree.c does, move the conditional branch
behind cell_content_area_offset < unallocated_space_offset validation.

We don't need to cache usable_minus_4 as BtreeContext because minus
instruction has small overhead enough compared to the whole n_local()
calculation.

This contains minor refactoring to change the pagesize type to u32 since
the max size is u16::MAX + 1.

+++ src/lib.rs
@@ -137,8 +137,10 @@ impl<'a> DatabaseHeader<'a> {
         self.0[20]
     }
 
-    pub fn usable_size(&self) -> i32 {
-        self.pagesize() as i32 - self.reserved() as i32
+    pub fn usable_size(&self) -> u32 {
+        // pagesize is bigger than or equal to 512.
+        // reserved is smaller than or equal to 255.
+        self.pagesize() - self.reserved() as u32
     }
 }
 
@@ -172,7 +174,7 @@ impl Connection {
         } else if !header.validate_reserved() {
             bail!("invalid reserved");
         }
-        let pager = Pager::new(file, header.pagesize() as usize)?;
+        let pager = Pager::new(file, header.pagesize())?;
         Ok(Self {
             pager,
             btree_ctx: BtreeContext::new(header.usable_size()),
