+++ src/test_utils.rs
@@ -43,10 +43,10 @@ pub fn create_pager(file: File) -> anyhow::Result<Pager> {
     let mut header_buf = [0_u8; DATABASE_HEADER_SIZE];
     file.read_exact_at(&mut header_buf, 0)?;
     let header = DatabaseHeader::from(&header_buf);
-    Pager::new(file, header.pagesize() as usize)
+    Pager::new(file, header.pagesize())
 }
 
-pub fn create_empty_pager(file_content: &[u8], pagesize: usize) -> Pager {
+pub fn create_empty_pager(file_content: &[u8], pagesize: u32) -> Pager {
     let file = NamedTempFile::new().unwrap();
     file.as_file().write_all_at(file_content, 0).unwrap();
     Pager::new(file.as_file().try_clone().unwrap(), pagesize).unwrap()
